% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-like.R
\name{add_prop_test}
\alias{add_prop_test}
\title{Add results of prop.test to data frame}
\usage{
add_prop_test(
  .df,
  x,
  n,
  p = NULL,
  CI_lower_name = "CI_lower",
  CI_upper_name = "CI_upper",
  p_name = "p",
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  correct = TRUE
)
}
\arguments{
\item{.df}{A data frame}

\item{x}{The column/vector with the number of positive results}

\item{n}{The column/vector/constant with the number of trials}

\item{p}{Assumed proportion: Will add a p-value that the proportion is equal to p (default: 0.5)}

\item{CI_lower_name, CI_upper_name, p_name}{Column names of the added columns}

\item{alternative, conf.level, correct}{As for \code{\link{prop.test}}}
}
\value{
Data frame with columns added
}
\description{
Adds prop.test results as columns to data frame based on data in columns
For use with a tibble in a pipe:
Using one-group prop.test, adds confidence intervals (with given conf.level)
for the proportion of x positive results in n trials,
and the p value that the proportion is equal to p (default: 0.5)
(to add the estimated proportion itself, use \code{\link{count_by}})
}
\examples{
library(magrittr)
if (requireNamespace("survival", quietly = TRUE))
{
   survival::aml \%>\%
   count_by(x) \%>\%
   add_prop_test(n, sum(n), rel)
}
}
\seealso{
\code{\link{count_by}()}
}
