% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-like.R
\name{add_summary_by}
\alias{add_summary_by}
\title{Add summary to tibble}
\usage{
add_summary_by(.data, .language, ...)
}
\arguments{
\item{.data}{Data argument, typical "first" argument in dplyr verbs}

\item{.language}{A call to \code{summarise()},
or another method performing equivalent aggregation (potentially wrapping \code{summarise()})}

\item{...}{Parameters for group_by}
}
\value{
The tibble with added columns
}
\description{
A verb for a dplyr pipeline:
Groups the frame by ... in addition to the current grouping,
then calls \code{\link{add_summary}}, then returns the frame with the mutated summarising columns
in the same grouping state as it was before this function was called.
}
