% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{all_or_all_na}
\alias{all_or_all_na}
\title{All() giving NA only if all values are NA}
\usage{
all_or_all_na(...)
}
\arguments{
\item{...}{Values}
}
\value{
NA if and only if all ... are NA, else all(...), ignoring NA values
}
\description{
All() giving NA only if all values are NA
}
