% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{any_or_all_na}
\alias{any_or_all_na}
\title{Any() giving NA only if all values are NA}
\usage{
any_or_all_na(...)
}
\arguments{
\item{...}{Values}
}
\value{
NA if and only if all ... are NA, else any(...), ignoring NA values
}
\description{
Any() giving NA only if all values are NA
}
