% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{are_true}
\alias{are_true}
\title{Vectorised conversion to logical, treating NA as False}
\usage{
are_true(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A logical vector of same size as x which is true where x is true (\code{\link[rlang]{as_logical}}) and not NA
}
\description{
Vectorised conversion to logical, treating NA as False
}
