% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr-like.R
\name{as_percentage_label}
\alias{as_percentage_label}
\title{Format as percentage for output}
\usage{
as_percentage_label(x, decimal_places = 1, include_plus_sign = F)
}
\arguments{
\item{x}{Numeric vector}

\item{decimal_places}{Decimal places to display}

\item{include_plus_sign}{prepend a "+" to the output if positive (if negative, a "-" must be prepended of course)}
}
\value{
Character vector
}
\description{
Vectorised conversion
}
\examples{
as_percentage_label(0.746) # gives "74.6\%"
}
