% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-like.R
\name{contingency_table_as_matrix}
\alias{contingency_table_as_matrix}
\title{Convert contingency table to classical R matrix}
\usage{
contingency_table_as_matrix(table_frame)
}
\arguments{
\item{table_frame}{Result of \code{\link{contingency_table_by}}}
}
\value{
A matrix
}
\description{
Converts the result of \code{\link{contingency_table_by}}
to a classical matrix
}
