% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dina.R
\name{dina}
\alias{dina}
\alias{dinAFormat}
\alias{dinA_format}
\alias{dinA}
\alias{dinAWidth}
\alias{dinA_width}
\alias{dinAHeight}
\alias{dinA_height}
\title{The DIN A paper formats}
\usage{
dinAFormat()

dinA_format()

dinA(n)

dinAWidth(n)

dinA_width(n)

dinAHeight(n)

dinA_height(n)
}
\arguments{
\item{n}{DIN A paper format index (0-10)}
}
\value{
A named list (0-10) of named vectors (long, short) of unit objects with the size in inches of the DIN A paper formats

named unit vector (long, short) with the size in inches of the requested DIN A paper format

the long side / width in landscape as a unit object in inches

the short side / height in landscape as a unit object in inches
}
\description{
The DIN A paper formats
}
\seealso{
\link[grid]{unit}
}
