% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{equal_including_na}
\alias{equal_including_na}
\title{Compare vectors, treating NA like a value}
\usage{
equal_including_na(v1, v2)
}
\arguments{
\item{v1, v2}{Vectors of equal size}
}
\value{
Returns a logical vector of the same size as v1 and v2, TRUE wherever elements are the same.
        NA is treated like a value level, i.e., NA == NA is true, NA == 1 is false.
}
\description{
Compare vectors, treating NA like a value
}
