% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-like.R
\name{execute_in_pipeline}
\alias{execute_in_pipeline}
\title{Executing language as if in a pipeline}
\usage{
execute_in_pipeline(.data, .language)
}
\arguments{
\item{.data}{A data frame}

\item{.language}{Language}
}
\value{
Result of the executed language
}
\description{
Executes the given language as if it was part of a magrittr pipeline\cr
\code{... \%>\% .language}\cr
while .data is the lhs value provided to .language as parameter by magrittr.
}
\details{
Note that language is evaluated as a quosure in its captured environment.
This is fine if this method is called as a secondary helper and .language is already a quosure;
otherwise you may want to explicitly set the quosure's environment to your caller's env.
}
