% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlang-like.R
\name{expression_list}
\alias{expression_list}
\alias{quosure_list}
\alias{symbol_string_list}
\title{Extract symbols from an expression of symbols and operators}
\usage{
expression_list(expr, seps = "+")

quosure_list(expr, seps = "+", env = caller_env())

symbol_string_list(expr, seps = "+")
}
\arguments{
\item{expr}{A language expression}

\item{seps}{Operators to consider as separators}

\item{env}{Environment for the created quosure}
}
\value{
A list of all symbols in the expression, as symbol, quosure or text.
}
\description{
Extract symbols from an expression of symbols and operators
}
\examples{
expression_list(a+b+c+d)
}
