% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{first_non_nas_at}
\alias{first_non_nas_at}
\title{Row-wise first value that is not NA}
\usage{
first_non_nas_at(.tbl, ...)
}
\arguments{
\item{.tbl}{A data frame}

\item{...}{A column selection, as for \code{\link[dplyr]{select}}}
}
\value{
A vector of length nrow(.tbl) containing the first found non-na value
}
\description{
Row-wise first value that is not NA
}
