% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{first_not}
\alias{first_not}
\title{First argument that does not equal a given value}
\usage{
first_not(not, ...)
}
\arguments{
\item{not}{Value: we look for the first value not equal to this one}

\item{...}{Values}
}
\value{
The first value that does not equal "not", or NA iff all equal "not"
}
\description{
First argument that does not equal a given value
}
\examples{
# 5
first_not(1, 1,1,1,5)
}
