% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{first_not_na}
\alias{first_not_na}
\title{First argument that is not NA}
\usage{
first_not_na(...)
}
\arguments{
\item{...}{Values}
}
\value{
The first argument that is not NA, or NA iff all are NA
}
\description{
First argument that is not NA
}
