% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{first_which_non_na_at}
\alias{first_which_non_na_at}
\title{Row-wise first index of column that is not NA}
\usage{
first_which_non_na_at(.tbl, ...)
}
\arguments{
\item{.tbl}{A data frame}

\item{...}{A column selection, as for \code{\link[dplyr]{select}}}
}
\value{
A numeric vector of length nrow(.tbl) containing the index of the first found non-na value in the given columns.
   Possible values are NA (all values in that row are NA), and 1 ... number of columns in selection
}
\description{
Row-wise first index of column that is not NA
}
