% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{first_which_not_na}
\alias{first_which_not_na}
\title{First which() is not na}
\usage{
first_which_not_na(...)
}
\arguments{
\item{...}{Values; concatenated as given. Intended use is with one vector of length > 1 or multiple single arguments.}
}
\value{
The index of the first value which is not NA, or NA iff all elements are NA.
}
\description{
First which() is not na
}
\examples{
# 4
first_which_not_na(NA, NA, NA, 56)
}
