% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{invalid}
\alias{invalid}
\alias{valid}
\title{A notion of valid and invalid}
\usage{
invalid(x)

valid(x)
}
\arguments{
\item{x}{Any object, value or NULL}
}
\value{
logical
}
\description{
An object is valid if it is not null, not missing (NA), and is not an empty vector.
Note that this is per se not vectorised, because a non-empty list or vector is valid as such.
}
\section{Functions}{
\itemize{
\item \code{valid()}: x is not invalid

}}
\examples{
invalid(NULL) # TRUE
invalid(NA) # TRUE
invalid(list()) # TRUE
invalid("a") # FALSE
invalid(c(1,2,3)) # FALSE
}
