% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr-like.R
\name{invert_value_and_names}
\alias{invert_value_and_names}
\title{Inverting name and value}
\usage{
invert_value_and_names(v)
}
\arguments{
\item{v}{A named vector}
}
\value{
A vector where names(v) are the values and the values of v are the names
}
\description{
Inverting name and value
}
