% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr-like.R
\name{lookup_function_from_dict}
\alias{lookup_function_from_dict}
\title{Creating a lookup function from dictionary}
\usage{
lookup_function_from_dict(dict, default = identity, dict_key_is_regex = F)
}
\arguments{
\item{dict}{A dictionaryish character vector (named: key -> value)}

\item{default}{Value to return if key is not found, or function to evaluate with key as argument}

\item{dict_key_is_regex}{If True, treats dictionary keys are regular expressions when matching}
}
\value{
A function which can be called with keys and performs the described lookup, returning the value (string)
}
\description{
Creating a lookup function from dictionary
}
