% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{named_palette}
\alias{named_palette}
\title{Named color palette}
\usage{
named_palette(palette, names, color_order = NULL)
}
\arguments{
\item{palette}{Colors}

\item{names}{Names}

\item{color_order}{If specified, will reorder palette by this ordering vector}
}
\value{
A named palette.
        If the palette is longer than names, will only use the first n entries.
        If names is longer than palette, will recycle colors.
}
\description{
Returns the palette named by names.
This is useful to pick only a few specific colors from a larger palette.
}
