% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcats-like.R
\name{order_factor_by}
\alias{order_factor_by}
\title{Reorder a factor}
\usage{
order_factor_by(.f, ...)
}
\arguments{
\item{.f}{A factor}

\item{...}{Passed to  \code{\link{order}()}. Should be vectors of the same size as .f.}
}
\value{
Reordered factor
}
\description{
Makes f a factor ordered according to ... (which is passed to order)
}
\details{
This is a thin wrapper around \code{\link[forcats]{fct_reorder}()}, which is unintuitive in conjunction with order().
}
\seealso{
\code{\link{rename_reorder_factor}}, \code{\link{rename_factor}}, \code{\link[forcats]{fct_reorder}}
}
