% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr-like.R
\name{orderer_function_from_sorted_vectors}
\alias{orderer_function_from_sorted_vectors}
\title{Orderer function for complex sorting}
\usage{
orderer_function_from_sorted_vectors(...)
}
\arguments{
\item{...}{k sorted vectors, in order of priority}
}
\value{
A function which takes (at least) k vectors
        This function will return an order for these vectors determined by the sorted vectors
}
\description{
If you want to order by multiple features and have sorted vectors for each feature
which describe the intended order
}
