% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr-like.R
\name{pluck_vector}
\alias{pluck_vector}
\title{Pluck with simplified return value}
\usage{
pluck_vector(.x, ..., .default = NULL)
}
\arguments{
\item{.x}{Container object}

\item{...}{Accessor specification}

\item{.default}{Default value}
}
\value{
Result of \code{\link[purrr]{pluck}()}, transformed y \code{\link[purrr]{simplify}()}
}
\description{
Like \code{\link[purrr]{pluck}()}, but will return simplify()'ed as a vector
}
