% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{prepare_directory}
\alias{prepare_directory}
\title{Directory creation}
\usage{
prepare_directory(folder)
}
\arguments{
\item{folder}{Folder path}
}
\value{
Folder path
}
\description{
Creates directory if it does not yet exist
}
