% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcats-like.R
\name{rename_factor}
\alias{rename_factor}
\title{Rename a factor.}
\usage{
rename_factor(.f, ..., reorder = F)
}
\arguments{
\item{.f}{A factor or vector (if .f is not yet a factor, it is made one)}

\item{...}{Dictionaryish arguments, named by old level, value is new level ("old level" = "new level").
You can pass single named arguments, or named vectors or named lists, which will be spliced.}

\item{reorder}{Logical: If True, the levels will additionally be reordered
in the order of first appearance in the arguments}
}
\value{
A renamed and reordered factor
}
\description{
Renames the levels of a factor.
}
\seealso{
\code{\link{rename_reorder_factor}}, \code{\link{order_factor_by}},
   \code{\link[forcats]{fct_recode}}, \code{\link[forcats]{fct_relevel}}
}
