% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr-like.R
\name{replace_sequential_duplicates}
\alias{replace_sequential_duplicates}
\title{Replace sequential duplicates}
\usage{
replace_sequential_duplicates(strings, replace_with = "", ordering = NULL)
}
\arguments{
\item{strings}{Character vector}

\item{replace_with}{Replacement string}

\item{ordering}{Optional: treat strings as if ordered like strings[ordering], or,
if a function, strings[ordering(strings)]}
}
\value{
A character vector with strings identical to the previous string replaced with replace_with
}
\description{
Replace sequential duplicates
}
\examples{
# returns c("a", "", "b", "", "", "a")
replace_sequential_duplicates(c("a", "a", "b", "b", "b", "a"))
}
