% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{save_pdf}
\alias{save_pdf}
\title{Save plot as PDF}
\usage{
save_pdf(plot, folder, fileBaseName, width, height, ...)
}
\arguments{
\item{plot}{A plot object that can be printed, e.g. result of ggplot2, plot_grid}

\item{folder}{Destination folder (will be created if it does not exist)}

\item{fileBaseName}{File base name (suffix ".pdf" will be added)}

\item{width, height}{PDF width and height in inches or as \code{grid::\link{unit}}.
If missing and the plot object has a "papersize" attribute c(width, height), this will be used.}

\item{...}{Further arguments which will be passed to \code{\link{cairo_pdf}}, e.g. family}
}
\description{
Save plot as PDF
}
