% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{save_png}
\alias{save_png}
\title{Save plot as PNG}
\usage{
save_png(
  plot,
  folder,
  fileBaseName,
  width,
  height,
  dpi = 300,
  background = c("white", "transparent"),
  ...
)
}
\arguments{
\item{plot}{A plot object that can be printed, e.g. result of ggplot2, plot_grid}

\item{folder}{Destination folder (will be created if it does not exist)}

\item{fileBaseName}{File base name (suffix ".png" will be added)}

\item{width, height}{PNG width and height in inches or as \code{grid::\link{unit}}.
If missing and the plot object has a "papersize" attribute c(width, height), this will be used.}

\item{dpi}{Resolution (determines file size in pixels, as size is given in inches)}

\item{background}{Initial background color, "white" or "transparent"}

\item{...}{Further arguments which will be passed to \code{\link{png}}, e.g. family}
}
\value{
invisible NULL
}
\description{
Save plot as PNG
}
