% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr-like.R
\name{sequential_duplicates}
\alias{sequential_duplicates}
\title{Detect sequential duplicates}
\usage{
sequential_duplicates(strings, ordering = NULL)
}
\arguments{
\item{strings}{Character vector}

\item{ordering}{Optional: treat strings as if ordered like strings[ordering], or,
if a function, strings[ordering(strings)]}
}
\value{
A logical vector which indicates if a string is identical to the previous string.
}
\description{
Detect sequential duplicates
}
\examples{
# return c(F, T, F, T, T, F)
sequential_duplicates(c("a", "a", "b", "b", "b", "a"))
}
