% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr-like.R
\name{str_locate_match}
\alias{str_locate_match}
\title{Combine str_match and str_locate}
\usage{
str_locate_match(patterns, strings)
}
\arguments{
\item{patterns}{Character vector of patterns}

\item{strings}{Character vector of strings}
}
\value{
Integer vector of length(patterns) where entry i gives the index in strings where pattern i first matched
}
\description{
For every pattern, return the index of the first match of pattern in strings
}
