% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlang-like.R
\name{syntactically_safe}
\alias{syntactically_safe}
\title{Syntactically safe names}
\usage{
syntactically_safe(expr_strings)
}
\arguments{
\item{expr_strings}{Strings to convert to syntactically safe form}
}
\value{
Strings converted to syntactically safe form
}
\description{
Makes the names syntactically safe by wrapping them in `` if necessary
}
