% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{true_or_na}
\alias{true_or_na}
\title{Test for logical true or NA}
\usage{
true_or_na(x)
}
\arguments{
\item{x}{Logical}
}
\value{
True if and only if x is TRUE or x is NA, False otherwise.
}
\description{
Test for logical true or NA
}
