% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{truthy}
\alias{truthy}
\alias{falsy}
\title{A python / javascript-like "truthy" notion}
\usage{
truthy(x)

falsy(x)
}
\arguments{
\item{x}{Any object, value or NULL}
}
\value{
logical
}
\description{
Values are truthy that are not null, NA, empty, 0, or FALSE.
}
\details{
Note that this is per se not vectorised, because a non-empty list or vector is "truthy" as such.
}
\section{Functions}{
\itemize{
\item \code{falsy()}: x is not truthy

}}
