% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuple-assignment.R
\name{tuple_assignment}
\alias{tuple_assignment}
\alias{\%=\%}
\alias{g}
\title{Infix operator for python-style tuple assignment}
\usage{
l \%=\% r

g(...)
}
\arguments{
\item{l}{left-hand side: "tuple" or variables created by \code{g()}}

\item{r}{right-hand side: Vector to assign to left-hand side variable}

\item{...}{Left-hand side variables to group}
}
\value{
Last assigned value
}
\description{
Infix operator for python-style tuple assignment
}
\examples{
g(a,b) \%=\% c(1,2) # equivalent to a <- 1; b <- 2
}
