% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr-like.R
\name{with_name}
\alias{with_name}
\alias{with_name_containing}
\alias{named}
\alias{name_contains}
\title{Slice by name}
\usage{
with_name(v, name)

with_name_containing(v, pattern)

named(name)

name_contains(pattern)
}
\arguments{
\item{v}{A vector}

\item{name}{Name of entry to pluck}

\item{pattern}{Pattern passed to \code{\link[stringr]{str_detect}}}
}
\value{
A slice from v containing all elements in v with the given name,
        or the name of which contains pattern
}
\description{
Slices of a vector with elements of given name, or containing given patterns.
Analogous accessor functions for \code{\link[purrr]{pluck}}
}
