% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr-like.R
\name{with_value_containing}
\alias{with_value_containing}
\alias{value_contains}
\title{Slice by value}
\usage{
with_value_containing(v, pattern)

value_contains(pattern)
}
\arguments{
\item{v}{A vector}

\item{pattern}{Pattern passed to \code{\link[stringr]{str_detect}}}
}
\value{
A slice from v containing all elements in v with the given name,
        or the name of which contains pattern
}
\description{
Slices of a vector with elements containing given patterns.
Analogous accessor function for \code{\link[purrr]{pluck}}
}
